Add files via upload
Dependency matrix is widely utilized in fault diagnosis. However, the diagnostic efficiency will decrease significantly with the matrix size increasing. Dependency matrix-based fault diagnosis can be regared as a searching process. Therefore, we introduce the hash-based searching and binary tree searching in the computer field to improve the diangosis efficiency. The codes is implemented in Matlab, and the running time is obtained for different methods.